requirements.txt
ipython
google-generativeai
python-dotenv
pandas
requests
beautifulsoup4
matplotlib
seaborn
papermill